[REF] - Final touch, refactor code, can be merged into main branch
