+++ CareerPathLog/View/LogScreenView.swift
